requirements.txt
Flask==0.10.1
Flask-Classy==0.6.10
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Jinja2==2.7.2
MarkupSafe==0.23
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.7.3
itsdangerous==0.24
nose==1.3.3
urllib3==1.8.2
wsgiref==0.1.2
